Update PacMangame.pde
Changed the initialisation of the ghost object. With the merge from Hannah's branch, "Pathfinding ghosts" i changed the initialisation of ghost to be on the same line so it just said "Ghost ghost = new Ghost()", but processing crashed when i did this. I am changing it back to what she has done, since that worked :)